// interfaces/listing.ts
export interface listing {
    id: string;
    url: string;
    scrape_id: string;
    medium_url: string;
    name: string;
    review_scores_rating: number;
    room_type: string;
    price: number;
    xl_picture_url: string;
    smart_location: string;
    number_of_reviews: number;
    number: string;
    bathrooms: number;
    campus: string;
    description: string;
    email: string;
   
}